requirements/requirements-common.txt
django-localflavor>=1.1
django-databrowse>=2014.9.26
django-queryset-csv>=0.3.1
